feat: refactor validation logic and reorganize repository interfaces
